adicionando logica

+++ src/main/java/com/example/backend_lp3/api/dto/PedidoDTO.java
@@ -17,7 +17,7 @@ public class PedidoDTO {
 
     private LocalDateTime dataPedido;
     private LocalDateTime dataEntrega;
-    private float precoTotal;
+    private Double precoTotal;
     private Long idFornecedor;
     private Long idGerente;
